feat: `ensureSignedInMiddleware()`
