+++ routes/account/upgrade.ts
@@ -1,14 +1,14 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import type { Handlers } from "$fresh/server.ts";
 import { stripe } from "@/utils/payments.ts";
-import type { AccountState } from "./_middleware.ts";
+import type { SignedInState } from "@/utils/middleware.ts";
 import { redirect } from "@/utils/redirect.ts";
 
 const STRIPE_PREMIUM_PLAN_PRICE_ID = Deno.env.get(
   "STRIPE_PREMIUM_PLAN_PRICE_ID",
 );
 
-export const handler: Handlers<null, AccountState> = {
+export const handler: Handlers<null, SignedInState> = {
   async GET(req, ctx) {
     if (
       !STRIPE_PREMIUM_PLAN_PRICE_ID || !ctx.state.sessionId ||
